requirements.txt
cffi==1.16.0
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
websockets==12.0
